Break down TestClientSteps into a set of related step definition files, that use StepsBase to access common properties.
