+++ test/AdaptiveRemote.EndToEndTests.Steps/Backend/HealthResponseSteps.cs
@@ -1,0 +1,12 @@
+﻿using System.Text.Json.Serialization.Metadata;
+using AdaptiveRemote.Contracts;
+using Reqnroll;
+
+namespace AdaptiveRemote.EndToEndTests.Steps.Backend;
+
+[Binding]
+internal static class HealthResponseSteps
+{
+    [StepArgumentTransformation(nameof(HealthResponse))]
+    public static JsonTypeInfo HealthResponseToJsonTypeInfo() => LayoutContractsJsonContext.Default.HealthResponse;
+}
