+++ test/AdaptiveRemote.EndtoEndTests.TestServices/TestClient.cs
@@ -1,6 +1,9 @@
 ﻿using System.Net.Http.Headers;
+using System.Text.Json;
+using System.Text.Json.Serialization.Metadata;
 using AdaptiveRemote.TestUtilities;
 using Microsoft.Extensions.Logging;
+using Microsoft.VisualStudio.TestTools.UnitTesting;
 
 namespace AdaptiveRemote.EndToEndTests.TestServices;
 
@@ -15,10 +18,21 @@ public class TestClient
     private readonly ILogger<TestClient> _log;
     private int _requestCount = 0;
 
+    private HttpResponseMessage? _lastResponseMessage;
+    private string? _lastResponseBody;
+    private object? _lastParsedObject = null;
+
     public TestClient(ILoggerFactory loggerFactory)
     {
         _log = loggerFactory.CreateLogger<TestClient>();
     }
+
+    public HttpResponseMessage LastResponse => _lastResponseMessage 
+        ?? throw new AssertFailedException("No request has been sent yet.");
+    public string LastResponseBody => _lastResponseBody
+        ?? throw new AssertFailedException("No request has been sent yet.");
+    public object LastResponseObject => _lastParsedObject 
+        ?? throw new AssertFailedException("The response body has not been deserialized yet. Ensure that the step 'the response body represents a {JsonTypeInfo}' is called before this step.");
 
     public HttpResponseMessage? SendRequest(HttpMethod method, Uri url, string? body = null)
     {
@@ -48,7 +62,9 @@ public class TestClient
             request.Headers.Authorization = new AuthenticationHeaderValue("Bearer", AuthorizationToken);
         }
 
-        HttpResponseMessage response = WaitHelpers.WaitForAsyncTask(ct => _httpClient.SendAsync(request, ct));
+        _lastParsedObject = null;
+        _lastResponseMessage = WaitHelpers.WaitForAsyncTask(ct => _httpClient.SendAsync(request, ct));
+        _lastResponseBody = WaitHelpers.WaitForAsyncTask(LastResponse.Content.ReadAsStringAsync);
 
         _log.LogInformation(
             """
@@ -58,11 +74,27 @@ public class TestClient
             """,
             _clientID,
             requestNumber,
-            (int)response.StatusCode,
-            response.ReasonPhrase,
-            response.Content.ReadAsStringAsync().Result);
+            (int)LastResponse.StatusCode,
+            LastResponse.ReasonPhrase,
+            LastResponse.Content.ReadAsStringAsync().Result);
 
-        return response;
+        return LastResponse;
+    }
+
+    public void ParseResponseAs(JsonTypeInfo jsonTypeInfo)
+    {
+        Assert.IsNotNull(LastResponseBody, "No response body to parse. Make sure to call SendRequest first and that the response has a body.");
+
+        try
+        {
+            _lastParsedObject = JsonSerializer.Deserialize(LastResponseBody, jsonTypeInfo);
+            Assert.IsNotNull(_lastParsedObject, "Deserialization returned null. Response body may be empty or not match the expected format.");
+        }
+        catch (JsonException ex)
+        {
+            Assert.Fail("Failed to parse the response body as JSON. {0}", ex.Message);
+            throw;
+        }
     }
 
     public override string ToString() => $"Client {_clientID}";
